feat(pipelines): add taskpill component
fix(pipelines): add correct location, scaling, and anchors for expanded task group's pill labels. (#10)

fix(pipelines): add status prop to group labels

update import statement
