+++ packages/module/src/components/nodes/labels/index.ts
@@ -3,3 +3,4 @@ export { default as LabelBadge } from './LabelBadge';
 export { default as LabelContextMenu } from './LabelContextMenu';
 export { default as LabelIcon } from './LabelIcon';
 export { default as NodeLabel } from './NodeLabel';
+export type { NodeLabelProps } from './NodeLabel';
