+++ packages/module/src/pipelines/components/groups/DefaultTaskGroup.tsx
@@ -10,6 +10,7 @@ import DefaultTaskGroupCollapsed from './DefaultTaskGroupCollapsed';
 import DefaultTaskGroupExpanded from './DefaultTaskGroupExpanded';
 import { RunStatus } from '../../types';
 import { DEFAULT_SPACER_NODE_TYPE } from '../../const';
+import { TaskGroupPillLabelProps } from './TaskGroupPillLabel';
 
 export interface EdgeCreationTypes {
   spacerNodeType?: string;
@@ -62,6 +63,8 @@ export interface DefaultTaskGroupProps {
   truncateLength?: number;
   /** Space between the label and the group. Defaults to 17 */
   labelOffset?: number;
+  /** Label to show for the group, Defaults to NodeLabel, only applicable to expanded groups */
+  GroupLabelComponent?: React.FC<TaskGroupPillLabelProps>;
   /** Center the label on the edge, overrides the label offset, only applicable to expanded groups */
   centerLabelOnEdge?: boolean;
   /** The Icon class to show in the label, ignored when labelIcon is specified */
@@ -126,7 +129,6 @@ type PipelinesDefaultGroupInnerProps = Omit<DefaultTaskGroupProps, 'element'> & 
 
 const DefaultTaskGroupInner: React.FunctionComponent<PipelinesDefaultGroupInnerProps> = observer(
   ({
-    className,
     element,
     badge,
     onCollapseChange,
@@ -207,7 +209,6 @@ const DefaultTaskGroupInner: React.FunctionComponent<PipelinesDefaultGroupInnerP
     if (element.isCollapsed()) {
       return (
         <DefaultTaskGroupCollapsed
-          className={className}
           element={element}
           shadowCount={collapsedShadowCount}
           onCollapseChange={handleCollapse}
@@ -216,10 +217,7 @@ const DefaultTaskGroupInner: React.FunctionComponent<PipelinesDefaultGroupInnerP
         />
       );
     }
-    return (
-      //  TODO: Support status indicators on expanded state.
-      <DefaultTaskGroupExpanded className={className} element={element} onCollapseChange={handleCollapse} {...rest} />
-    );
+    return <DefaultTaskGroupExpanded element={element} badge={badge} onCollapseChange={handleCollapse} {...rest} />;
   }
 );
