+++ packages/module/src/pipelines/components/groups/index.ts
@@ -2,3 +2,4 @@ export type { EdgeCreationTypes } from './DefaultTaskGroup';
 export { default as DefaultTaskGroup } from './DefaultTaskGroup';
 export { default as DefaultTaskGroupExpanded } from './DefaultTaskGroupExpanded';
 export { default as DefaultTaskGroupCollapsed } from './DefaultTaskGroupCollapsed';
+export { default as TaskGroupPillLabel } from './TaskGroupPillLabel';
