+++ packages/module/src/pipelines/components/nodes/index.ts
@@ -2,4 +2,5 @@ export { default as FinallyNode } from './FinallyNode';
 export { default as SpacerNode } from './SpacerNode';
 export { default as StatusIcon } from '../../utils/StatusIcon';
 export { default as TaskNode } from './TaskNode';
+export { default as TaskPill } from './TaskPill';
 export { default as WhenNode } from '../../decorators/WhenDecorator';
